Restore int for timestamps

+++ POPSensorData.cc
@@ -41,8 +41,7 @@ RecordHeader::RecordHeader()
 	unit            = UNT_NONE;
 	timeStamp       = 0;
 }
-
-RecordHeader::RecordHeader(unsigned long x_timeStamp, const NotifyMessage& x_msg)
+RecordHeader::RecordHeader(unsigned int x_timeStamp, const NotifyMessage& x_msg)
 {
 	// cout << "Create record header id:" << x_msg.id << " units " << x_msg.unit << popcendl;
 	measurementType = x_msg.measurementType;
@@ -63,7 +62,7 @@ void RecordHeader::Serialize(POPBuffer &buf, bool pack)
 		int mu = static_cast<int>(unit);
 		buf.Pack(&mu,1);
 
-		unsigned long ts = timeStamp;
+		int ts = (int)timeStamp;
 		buf.Pack(&timeStamp,1);
 	}
 	else
@@ -185,7 +184,7 @@ void POPSensorData::ReadFromFile(istream& xr_istream)
 	// Read each line
 	while(getline(xr_istream,line))
 	{
-		unsigned long timeStamp;
+		unsigned int timeStamp;
 		RecordHeader rec;
 		stringstream lineStream(line);
 		string       cell;
